+++ src/pages/trade/api/book.ts
@@ -1,22 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
-import { Position } from '@penumbra-zone/protobuf/penumbra/core/component/dex/v1/dex_pb';
-import { JsonValue } from '@bufbuild/protobuf';
 import { useRefetchOnNewBlock } from '@/shared/api/compact-block.ts';
-import {
-  deserializeRouteBookResponseJson,
-  RouteBookResponse,
-  RouteBookResponseJson,
-} from '@/shared/api/server/booktwo.ts';
-
-interface BookResponse {
-  asks: Position[];
-  bids: Position[];
-}
-
-interface BookResponseJson {
-  asks: JsonValue[];
-  bids: JsonValue[];
-}
+import { RouteBookResponse, RouteBookResponseJson } from '@/shared/api/server/book/types';
+import { deserializeRouteBookResponseJson } from '@/shared/api/server/book/serialization.ts';
 
 export const useBook = (symbol1: string | undefined, symbol2: string | undefined) => {
   const query = useQuery({
@@ -30,7 +15,7 @@ export const useBook = (symbol1: string | undefined, symbol2: string | undefined
         baseAsset: symbol1,
         quoteAsset: symbol2,
       };
-      const baseUrl = '/api/booktwo';
+      const baseUrl = '/api/book';
       const urlParams = new URLSearchParams(paramsObj).toString();
       const res = await fetch(`${baseUrl}?${urlParams}`);
       const data = (await res.json()) as RouteBookResponseJson;
