+++ src/shared/api/server/book/types.ts
@@ -1,0 +1,38 @@
+import { JsonValue } from '@bufbuild/protobuf';
+import { ValueView } from '@penumbra-zone/protobuf/penumbra/core/asset/v1/asset_pb';
+
+export interface Trace {
+  price: string;
+  amount: string;
+  total: string;
+  hops: ValueView[];
+}
+
+export type TraceHash = string;
+
+export interface BuySellTraces {
+  buy: Trace[];
+  sell: Trace[];
+}
+
+export interface RouteBookResponse {
+  singleHops: BuySellTraces;
+  multiHops: BuySellTraces;
+}
+
+export interface TraceJson {
+  price: string;
+  amount: string;
+  total: string;
+  hops: JsonValue[];
+}
+
+export interface BuySellTracesJson {
+  buy: TraceJson[];
+  sell: TraceJson[];
+}
+
+export interface RouteBookResponseJson {
+  singleHops: BuySellTracesJson;
+  multiHops: BuySellTracesJson;
+}
